+++ mopidy/audio/utils.py
@@ -8,9 +8,8 @@ import pygst
 pygst.require('0.10')
 import gst  # noqa
 
-from mopidy import compat
+from mopidy import compat, httpclient
 from mopidy.models import Album, Artist, Track
-from mopidy.utils import http
 
 logger = logging.getLogger(__name__)
 
@@ -143,7 +142,7 @@ def setup_proxy(element, config):
     if not hasattr(element.props, 'proxy') or not config.get('hostname'):
         return
 
-    element.set_property('proxy', http.format_proxy(config, auth=False))
+    element.set_property('proxy', httpclient.format_proxy(config, auth=False))
     element.set_property('proxy-id', config.get('username'))
     element.set_property('proxy-pw', config.get('password'))